Updated Card.getFullSet to use flatMap and map over nested loops
Card.getFullSet() created a HashSet and then used nested for each
loops to add all card types to the set. This was a good use case for
flatMap and map over the provided lists, and a toHashSet function call.
Reduced the getFullSet function to a declarative one liner. Test output
for this was added to the main function.